+++ tests/testthat/test-size.R
@@ -157,6 +157,17 @@ test_that("na of list-array is 1d slice", {
   expect_equal(x2, array(list(), c(1, 3, 4)))
 })
 
+# sequences ---------------------------------------------------------------
+
+test_that("vec_seq_along returns size-0 output for size-0 input", {
+  expect_equal(vec_seq_along(character()), integer())
+  expect_equal(vec_seq_along(data.frame()), integer())
+})
+
+test_that("vec_na_along can be called with single argument", {
+  expect_equal(vec_na_along(1:3), rep(NA_integer_, 3))
+})
+
 # vec_names ---------------------------------------------------------------
 
 test_that("can retrieve names", {